requirements.txt
asgiref==3.7.2
beautifulsoup4==4.9.3
certifi==2023.7.22
colorama==0.4.4
DAWG-Python==0.7.2
delocate==0.10.4
dj-database-url==0.5.0
Django==4.2.5
django-contrib-comments==2.2.0
django-cors-headers==3.7.0
django-elasticsearch-dsl==8.0
django-environ==0.11.2
django-haystack==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
docopt==0.6.2
docutils==0.20.1
drf-yasg==1.21.7
elastic-transport==8.4.1
elasticsearch==8.10.1
elasticsearch-dsl==8.9.0
gunicorn==20.1.0
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
PyYAML==6.0.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.18
whitenoise==5.2.0
Whoosh==2.7.4
